Move model into rspotify-model
