+++ Cargo.toml
@@ -11,6 +11,8 @@ keywords = ["spotify","api"]
 edition = "2018"
 
 [dependencies]
+rspotify-model = { path = "./model", version = "0.10.0" }
+
 # Temporary until https://github.com/rust-lang/rfcs/issues/2739, for
 # `maybe_async`.
 async-trait = { version = "0.1.40", optional = true }
@@ -28,7 +30,6 @@ thiserror = "1.0.20"
 ureq = { version = "2.0", default-features = false, features = ["json", "cookies"], optional = true }
 url = "2.1.1"
 webbrowser = { version = "0.5.5", optional = true }
-strum = { version = "0.20", features = ["derive"] }
 
 [dev-dependencies]
 env_logger = "0.8.1"
