+++ model/src/album.rs
@@ -9,7 +9,7 @@ use super::image::Image;
 use super::page::Page;
 use super::track::SimplifiedTrack;
 use super::Restriction;
-use crate::model::{AlbumType, Copyright, DatePrecision, Type};
+use crate::{AlbumType, Copyright, DatePrecision, Type};
 
 /// Simplified Album Object
 ///
@@ -67,7 +67,7 @@ pub struct FullAlbum {
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-multiple-albums)
 #[derive(Deserialize)]
-pub(in crate) struct FullAlbums {
+pub struct FullAlbums {
     pub albums: Vec<FullAlbum>,
 }
 
@@ -75,7 +75,7 @@ pub(in crate) struct FullAlbums {
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-new-releases)
 #[derive(Deserialize)]
-pub(in crate) struct PageSimpliedAlbums {
+pub struct PageSimpliedAlbums {
     pub albums: Page<SimplifiedAlbum>,
 }
