+++ model/src/enums/country.rs
