+++ model/src/enums/misc.rs
