+++ model/src/enums/mod.rs
