+++ model/src/image.rs
