+++ model/src/playlist.rs
@@ -7,7 +7,7 @@ use super::image::Image;
 use super::page::Page;
 use super::track::FullTrack;
 use super::user::PublicUser;
-use crate::model::{Followers, Type};
+use crate::{Followers, Type};
 
 /// Playlist result object
 ///
@@ -90,6 +90,6 @@ pub struct FeaturedPlaylists {
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-a-categories-playlists)
 #[derive(Deserialize)]
-pub(in crate) struct CategoryPlaylists {
+pub struct CategoryPlaylists {
     pub playlists: Page<SimplifiedPlaylist>,
 }
