+++ model/src/recommend.rs
@@ -1,6 +1,6 @@
 //! All objects related to recommendation
 use super::track::SimplifiedTrack;
-use crate::model::RecommendationsSeedType;
+use crate::RecommendationsSeedType;
 use serde::{Deserialize, Serialize};
 
 /// Recommendations object
