+++ model/src/search.rs
