+++ model/src/show.rs
@@ -1,6 +1,6 @@
 use super::image::Image;
 use super::page::Page;
-use crate::model::{duration_ms, CopyrightType, DatePrecision};
+use crate::{duration_ms, CopyrightType, DatePrecision};
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::time::Duration;
@@ -42,7 +42,7 @@ pub struct SimplifiedShow {
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-multiple-shows)
 #[derive(Deserialize)]
-pub(in crate) struct SeversalSimplifiedShows {
+pub struct SeversalSimplifiedShows {
     pub shows: Vec<SimplifiedShow>,
 }
