+++ model/src/track.rs
@@ -7,7 +7,7 @@ use std::{collections::HashMap, time::Duration};
 use super::album::SimplifiedAlbum;
 use super::artist::SimplifiedArtist;
 use super::Restriction;
-use crate::model::{duration_ms, TrackId, Type};
+use crate::{duration_ms, TrackId, Type};
 
 /// Full track object
 ///
@@ -59,7 +59,7 @@ pub struct TrackLink {
 ///
 /// [Reference](https://developer.spotify.com/documentation/web-api/reference/#endpoint-get-several-tracks)
 #[derive(Deserialize)]
-pub(in crate) struct FullTracks {
+pub struct FullTracks {
     pub tracks: Vec<FullTrack>,
 }
